+++ CHANGELOG.md
@@ -17,6 +17,8 @@ and std:io:file:append.
 parent GlobalEnv.
 * Feature: Added SymbolTable::set() for setting variables that are provided by a module
 from Rust.
+* Feature: Added many num:\* functions (still undocumented)
+* Feature: Added hash:fnv1a, and rand:split\_mix64\_\* functions.
 
 0.3.1 (2019-09-16)
 ==================